pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "cloud_env_simulator"
version = "0.1.0"
description = "Cloud Infrastructure Incident Simulator"
authors = [{ name = "SRE AI" }]
dependencies = [
    "openenv-core>=0.2.0",
    "pydantic>=2.0.0",
    "fastapi",
    "uvicorn"
]

[project.scripts]
server = "server.app:main"
